requirements.txt
biopython==1.79
click==8.1.3
cloudpickle==2.1.0
colorama==0.4.5
dask==2022.1.1
distributed==2022.1.1
fsspec==2022.8.2
HeapDict==1.0.1
Jinja2==3.1.2
locket==1.0.0
MarkupSafe==2.1.1
modin==0.15.2
msgpack==1.0.4
numpy==1.23.2
packaging==21.3
pandas==1.4.3
partd==1.3.0
pgzip==0.3.2
psutil==5.9.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
sortedcontainers==2.4.0
tblib==1.7.0
toolz==0.12.0
tornado==6.2
zict==2.2.0
